+++ oauth2_provider/urls.py
@@ -8,3 +8,13 @@ urlpatterns = patterns(
     url(r'^authorize/$', views.AuthorizationView.as_view(), name="authorize"),
     url(r'^token/$', views.TokenView.as_view(), name="token"),
 )
+
+# Application management views
+urlpatterns += patterns(
+    '',
+    url(r'^applications/$', views.ApplicationList.as_view(), name="list"),
+    url(r'^applications/register/$', views.ApplicationRegistration.as_view(), name="register"),
+    url(r'^applications/(?P<pk>\d+)/$', views.ApplicationDetail.as_view(), name="detail"),
+    url(r'^applications/(?P<pk>\d+)/delete/$', views.ApplicationDelete.as_view(), name="delete"),
+    url(r'^applications/(?P<pk>\d+)/update/$', views.ApplicationUpdate.as_view(), name="update"),
+)