+++ oauth2_provider/views/__init__.py
@@ -1,2 +1,4 @@
 from .base import AuthorizationView, TokenView
+from .application import ApplicationRegistration, ApplicationDetail, ApplicationList, \
+    ApplicationDelete, ApplicationUpdate
 from .generic import ProtectedResourceView, ScopedProtectedResourceView, ReadWriteScopedResourceView